+++ apps/storybook/src/stories/3 - ui/Button.stories.tsx
@@ -1,3 +1,4 @@
+import React from 'react'
 import { StoryFn, Meta } from '@storybook/react'
 import { fn } from '@storybook/test'
 import {