+++ platforms/nextjs/src/theme/index.ts
@@ -1,5 +1,5 @@
-import { webVariants } from '@shallot-ui/platform-react'
+import { webVariants, reactThemeModes } from '@shallot-ui/platform-react'
 import { makeTheme } from '@shallot-ui/core-theme'
 
-export const DEFAULT_NEXT_THEME = makeTheme({}, webVariants)
+export const DEFAULT_NEXT_THEME = makeTheme({}, webVariants, reactThemeModes)
 export type NextTheme = typeof DEFAULT_NEXT_THEME