+++ platforms/react/src/theme/index.ts
@@ -1,5 +1,6 @@
 import { makeTheme } from '@shallot-ui/core-theme'
 import webVariants from './variants'
+import reactThemeModes from './modes'
 
-export { webVariants }
-export const DEFAULT_WEB_THEME = makeTheme({}, webVariants)
+export { webVariants, reactThemeModes }
+export const DEFAULT_WEB_THEME = makeTheme({}, webVariants, reactThemeModes)